requirements.txt
blinker==1.8.2
certifi==2026.2.25
cffi==1.17.1
charset-normalizer==3.4.5
click==8.1.8
colorama==0.4.6
cryptography==38.0.4
dnspython==2.6.1
email-validator==2.3.0
et_xmlfile==2.0.0
Flask==3.0.3
Flask-Login==0.6.3
Flask-WTF==1.2.1
idna==3.11
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==2.1.5
numpy==1.24.4
openpyxl==3.1.5
packaging==26.0
pandas==2.0.3
passlib==1.7.4
pdfminer.six==20221105
pdfplumber==0.9.0
pillow==10.4.0
pycparser==2.23
pypdfium2==5.6.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.4
six==1.17.0
typing_extensions==4.13.2
tzdata==2025.3
urllib3==2.2.3
Wand==0.7.0
Werkzeug==3.0.6
WTForms==3.1.2
zipp==3.20.2
